+++ packages/luciq_flutter_modular/example/lib/main.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_modular_demo_app/modules.dart';
@@ -7,19 +5,11 @@ import 'package:luciq_flutter/luciq_flutter.dart';
 import 'package:luciq_flutter_modular/luciq_flutter_modular.dart';
 
 void main() {
-  runZonedGuarded(
-    () {
-      WidgetsFlutterBinding.ensureInitialized();
-
-      Luciq.init(
-        token: '6b41bc30dd42aac50794ef3ec8f74a74',
-        invocationEvents: [InvocationEvent.floatingButton],
-        debugLogsLevel: LogLevel.verbose,
-      );
-
-      FlutterError.onError = (FlutterErrorDetails details) {
-        Zone.current.handleUncaughtError(details.exception, details.stack!);
-      };
+  Luciq.init(
+    token: '6b41bc30dd42aac50794ef3ec8f74a74',
+    invocationEvents: [InvocationEvent.floatingButton],
+    debugLogsLevel: LogLevel.verbose,
+    appRunner: () {
       runApp(
         ModularApp(
           module: LuciqModule(AppModule()),
@@ -27,7 +17,6 @@ void main() {
         ),
       );
     },
-    CrashReporting.reportCrash,
   );
 }